% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{acc_met}
\alias{acc_met}
\title{Calculate accuracy metrics for interval data}
\usage{
acc_met(cent_pred, cent_act, ran_pred, ran_act, train_resp)
}
\arguments{
\item{cent_pred}{A data frame of the center values of the predictions
from the testing data.}

\item{cent_act}{A data frame of the center values of the actual
from the testing data.}

\item{ran_pred}{A data frame of the range values of the predictions
from the testing data.}

\item{ran_act}{A data frame of the range values of the actual
from the testing data.}

\item{train_resp}{A data frame of the outcome training values, used
for standardizing the accuracy metrics.}
}
\value{
A data frame with the calculated accuracy metrics including
mean squared error (MSE), mean absolute error (MAE), and R-squared (R2)
values for the intervals.
}
\description{
The function \code{acc_met} takes in predicted and actual center and range values
of interval data, along with the outcome training data, and calculates
the mean squared error (MSE), mean absolute error (MAE), and R-squared (R2) values
for the intervals.
}
\examples{
cent_pred <- data.frame(1, 2, 3)
cent_act <- data.frame(1.1, 1.8, 2.9)
ran_pred <- data.frame(0.2, 0.3, 0.1)
ran_act <- data.frame(0.3, 0.2, 0.15)
train_resp <- data.frame(1.1, 1.2, 1.3)
acc_met(cent_pred, cent_act, ran_pred, ran_act, train_resp)

}
