% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Int_plot.R
\name{int_plot}
\alias{int_plot}
\title{Plots Interaction between Two Interval-valued Variables}
\usage{
int_plot(int_data, title, xlabel, ylabel)
}
\arguments{
\item{int_data}{A data frame with 4 columns representing the centers and
ranges of the two variables. The columns should be named as "yc", "yr", "xc",
and "xr" respectively.}

\item{title}{A character string specifying the title of the plot.}

\item{xlabel}{A character string specifying the label of the x-axis.}

\item{ylabel}{A character string specifying the label of the y-axis.}
}
\value{
A ggplot object containing the interaction plot.
}
\description{
This function plots the interaction between two variables by creating a
bounding box and center points for the variables.
}
\examples{
int_data <- data.frame(yc = 0, yr = 1, xc = 0, xr = 1)
int_plot(int_data, "Interaction Plot", "X-axis", "Y-axis")

}
