% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF.R
\name{intrf}
\alias{intrf}
\title{Interval-valued Random Forests}
\usage{
intrf(
  int_resp,
  cent_pred,
  ran_pred,
  train,
  test = NULL,
  num_trees = 500,
  boot_perc = 0.8,
  mtry_int = NULL
)
}
\arguments{
\item{int_resp}{A data frame of interval-valued response variable.}

\item{cent_pred}{A data frame of the center values of the predictor
variables.}

\item{ran_pred}{A data frame of the range values of the predictor
variables.}

\item{train}{!IDK if I want this! A data frame of the training set of the
interval-valued data}

\item{test}{!IDK if I want this! A data frame of the testing set of the
interval-valued data (default = NULL)}

\item{num_trees}{Number of trees used for ensembling (default = 500).}

\item{boot_perc}{Percent of observations used in the bootstrapping process
(default = 0.8).}

\item{mtry_int}{Number of interval predictor variables sampled (default =
(p^c + p^r )/6)}
}
\value{
A random forest model with the bootstrapped information, individual
 trees and their predictions, and the interval predictions.
}
\description{
Interval-valued Random Forests
}
\examples{
# set seed for reproducibility
set.seed(123)

int_resp <- data.frame(low = runif(100), high = runif(100))

# create center and range values of predictor variables
cent_pred <- data.frame(x1 = runif(100), x2 = runif(100), x3 = runif(100))
ran_pred <- data.frame(x1 = runif(100), x2 = runif(100), x3 = runif(100))
ran_pred <- ran_pred - cent_pred

# create training set
train <- cbind(int_resp, cent_pred, ran_pred)

intrf(int_resp = int_resp, cent_pred = cent_pred,
ran_pred = ran_pred, train = train, num_trees = 100)

}
