% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{make_lu_int}
\alias{make_lu_int}
\title{Boxes for interval data}
\usage{
make_lu_int(int_data)
}
\arguments{
\item{int_data}{A data frame of center and range interval-valued data
(copy of int_plot).}
}
\description{
Boxes for interval data
}
