% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{pairsamp}
\alias{pairsamp}
\title{Pairwise Sampling (Helper Function)}
\usage{
pairsamp(int_resp, cent_pred, ran_pred, train, mtry)
}
\arguments{
\item{int_resp}{A data frame of interval-valued response variable
(copy of int_resp in intrf).}

\item{cent_pred}{A data frame of the center values of the predictor
variables (copy of cent_pred in intrf).}

\item{ran_pred}{A data frame of the range values of the predictor
variables (copy of ran_pred in intrf).}

\item{train}{A data frame of the training set of the
interval-valued data (copy of train in intrf).}

\item{mtry}{Number of interval predictor variables sampled (default =
(p^c + p^r )/6) (copy of mtry_int in intrf).}
}
\value{
a list containing the original data set and a subset of the predictors.
}
\description{
Pairwise Sampling (Helper Function)
}
\examples{

# Simulate data
set.seed(123)
int_resp <- rnorm(100)
cent_pred <- matrix(rnorm(500), nrow = 100)
ran_pred <- matrix(runif(500), nrow = 100)
train <- data.frame(int_resp, cent_pred, ran_pred)

# Randomly sample subset of predictors
mtry <- 3
pairsamp(int_resp, cent_pred, ran_pred, train, mtry)

}
