% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{prices_stand}
\alias{prices_stand}
\title{A data frame of stock prices for GE, where each variable is standardized
by its respective variance.}
\format{
A data frame with 1509 rows and 4 variables:
\describe{
\item{c.DJI}{Center values for The Dow Jones Industrial Average}

\item{r.DJI}{Range values for The Dow Jones Industrial Average}

\item{c.GE}{Center values for General Electric}

\item{r.GE}{Range values for General Electric}
}
}
\description{
A data frame of stock prices for GE, where each variable is standardized
by its respective variance.
}
